% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batsmanDismissals.R
\name{batsmanDismissals}
\alias{batsmanDismissals}
\title{Compute and plot the dismissal type for the batsman}
\usage{
batsman4s(df, name="Kohli")
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of batsman}
}
\value{
None
}
\description{
This function computes and plots the type of dismissals of the
the batsman
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get the data frame for Kohli
kohli <- getBatsmanDetails(team="India",name="Kohli")
batsmanDismissals(kohli,"Kohli")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{batsmanFoursSixes}}
\code{\link{batsmanRunsVsDeliveries}}
\code{\link{batsmanRunsVsStrikeRate}}
}

