% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batsmanFoursSixes.R
\name{batsmanFoursSixes}
\alias{batsmanFoursSixes}
\title{Compute and plot the total runs, fours and sixes of the batsman}
\usage{
batsmanFoursSixes(df, name="Kohli")
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of batsman}
}
\value{
None
}
\description{
This function computes and plots the total runs, fours and sixes of
the batsman
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get the data frame for Kohli
kohli <- getBatsmanDetails(team="India",name="Kohli")
kohli46 <- select(kohli,batsman,ballsPlayed,fours,sixes,runs)
batsmanFoursSixes(kohli46,"Kohli")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{batsmanDismissals}}
\code{\link{batsmanRunsVsDeliveries}}
\code{\link{batsmanRunsVsStrikeRate}}
\code{\link{batsmanRunsVsStrikeRate}}
\code{\link{batsmansRunsPredict}}
}

