% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batsmanMovingAverage.R
\name{batsmanMovingAverage}
\alias{batsmanMovingAverage}
\title{Compute and plot moving average of the batsman over the batsman's career}
\usage{
batsmanMovingAverage(df, name= "A Leg Glance")
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of batsman}
}
\value{
None
}
\description{
This function plots the runs scored by the batsman over the career as a time
series. A loess regression line is plotted on the moving average of the batsman
the batsman
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get the data frame for Kohli
kohli <- getBatsmanDetails(team="India",name="Kohli")
batsmanMovingAverage(kohli,"Kohli")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{batsmanDismissals}}
\code{\link{batsmanRunsVsDeliveries}}
\code{\link{batsmanRunsVsStrikeRate}}
\code{\link{batsmanRunsVsStrikeRate}}
\code{\link{batsmanRunsPredict}}
\code{\link{teamBatsmenPartnershipAllOppnAllMatches}}
}

