% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batsmanRunsAgainstOpposition.R
\name{batsmanRunsAgainstOpposition}
\alias{batsmanRunsAgainstOpposition}
\title{Compute and plot the average runs scored by the batsman against different oppositions}
\usage{
batsmanRunsAgainstOpposition(df, name="Kohli")
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of batsman}
}
\value{
None
}
\description{
This function computes and plots the mean runs scored by the batsman against different
oppositions
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get the data frame for Kohli
kohli <- getBatsmanDetails(team="India",name="Kohli")
batsmanRunsAgainstOpposition(kohli,"Kohli")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{batsmanFoursSixes}}
\code{\link{batsmanRunsVsDeliveries}}
\code{\link{batsmanRunsVsStrikeRate}}
\code{\link{batsmanRunsPredict}}
\code{\link{teamBatsmenPartnershipAllOppnAllMatches}}
}

