% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batsmanRunsVenue.R
\name{batsmanRunsVenue}
\alias{batsmanRunsVenue}
\title{Compute and plot the average runs scored by the batsman  at different venues}
\usage{
batsmanRunsVenue(df, name="Kohli")
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of batsman}
}
\value{
None
}
\description{
This function computes and plots the mean runs scored by the batsman at different
venues of the world
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get the data frame for Kohli
kohli <- getBatsmanDetails(team="India",name="Kohli")
batsmanRunsVenue(kohli,"Kohli")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{batsmanFoursSixes}}
\code{\link{batsmanRunsVsDeliveries}}
\code{\link{batsmanRunsVsStrikeRate}}
\code{\link{batsmansRunsPredict}}
\code{\link{teamBatsmanPartnershipAllOppnAllMatches}}
\code{\link{batsmanRunsAgainstOpposition}}
}

