% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batsmansRunsVsDeliveries.R
\name{batsmanRunsVsDeliveries}
\alias{batsmanRunsVsDeliveries}
\title{Runs versus deliveries faced}
\usage{
batsmanRunsVsDeliveries(df, name= "A Late Cut",dateRange,staticIntv=1)
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of batsman}

\item{dateRange}{Date interval to consider}

\item{staticIntv}{Static or interactive -staticIntv =1 (static plot) &  staticIntv =2 (interactive  plot)}
}
\value{
None
}
\description{
This function plots the runs scored and the deliveries required. A  regression
smoothing function is used to fit the points
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#Get the data frame for Kohli
kohli <- getBatsmanDetails(team="India",name="Kohli",dir=pathToFile)
batsmanRunsVsDeliveries(kohli,"Kohli")
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{batsmanFoursSixes}}\cr
\code{\link{batsmanRunsVsDeliveries}}\cr
\code{\link{batsmanRunsVsStrikeRate}}\cr
}
\author{
Tinniam V Ganesh
}
