% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlerDeliveryWickets.R
\name{bowlerDeliveryWickets}
\alias{bowlerDeliveryWickets}
\title{This function creates a dataframe of deliveries bowled and wickets taken}
\usage{
bowlerDeliveryWickets(match,theTeam,name)
}
\arguments{
\item{match}{Data frame of the match}

\item{theTeam}{The team for which the delivery wickets have to be computed}

\item{name}{The name of the bowler}
}
\value{
dataframe
}
\description{
This function creates a dataframe of balls bowled versus the wickets taken by
the bowler
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get the data frame for Kohli
kohli <- getBatsmanDetails(team="India",name="Kohli")
batsmanDismissals(kohli,"Kohli")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{batsmanFoursSixes}}
\code{\link{batsmanRunsVsDeliveries}}
\code{\link{batsmanRunsVsStrikeRate}}
\code{\link{bowlerDeliveryWickets}}
\code{\link{bowlerMeanEconomyRate}}
\code{\link{bowlerMeanRunsConceded}}
}

