% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlerMeanEconomyRate.R
\name{bowlerMeanEconomyRate}
\alias{bowlerMeanEconomyRate}
\title{Mean economy rate versus number of overs}
\usage{
bowlerMeanEconomyRate(df, name)
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of bowler}
}
\value{
None
}
\description{
This function computes and plots mean economy rate and the number of
overs bowled by the bowler
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
# Get the data frame for RA Jadeja
jadeja <- getBowlerWicketDetails(team="India",name="Jadeja",dir=pathToFile)
bowlerMeanEconomyRate(jadeja,"RA Jadeja")
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{bowlerMovingAverage}}\cr
\code{\link{bowlerWicketPlot}}\cr
\code{\link{bowlerWicketsVenue}}\cr
\code{\link{bowlerMeanRunsConceded}}\cr
}
\author{
Tinniam V Ganesh
}
