% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlerMovingAverage.R
\name{bowlerMovingAverage}
\alias{bowlerMovingAverage}
\title{Compute and plot the moving average of the wickets taken by bowler over career}
\usage{
bowlerMovingAverage(df, name)
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of bowler}
}
\value{
None
}
\description{
This function computes and plots the wickets taken by the bowler over career. A loess
regression fit plots the moving average of wickets taken by bowler
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
# Get the data frame for RA Jadeja
jadeja <- getBowlerWicketDetails(team="India",name="Jadeja",dir=pathToFile)
bowlerMeanRunsConceded(jadeja,"RA Jadeja")
}

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{bowlerMeanEconomyRate}}
\code{\link{bowlerWicketPlot}}
\code{\link{bowlerWicketsVenue}}
\code{\link{bowlerMeanRunsConceded}}
}

