% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlerVsBatsmanPerf.R
\name{bowlerVsBatsmanPerf}
\alias{bowlerVsBatsmanPerf}
\title{Performance of bowler vs batsman}
\usage{
bowlerVsBatsmanPerf(t20MDF,batsman1,bowler1)
}
\arguments{
\item{t20MDF}{Dataframe}

\item{batsman1}{The batsman}

\item{bowler1}{The bowler}
}
\value{
None
}
\description{
This function computes the performance of bowler vs batsman
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{

bowlerVsBatsmanPerf(t20MDF,batsman1,bowler1)
}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{batsmanFoursSixes}}\cr
\code{\link{batsmanRunsVsDeliveries}}\cr
\code{\link{batsmanRunsVsStrikeRate}}\cr
}
\author{
Tinniam V Ganesh
}
