% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlerWicketsVenue.R
\name{bowlerWicketsVenue}
\alias{bowlerWicketsVenue}
\title{Compute and plot mean wicket taken by the bowler at different venues}
\usage{
bowlerWicketsVenue(df, name)
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of bowler}
}
\value{
None
}
\description{
This function computes and plots mean number of wickets taken by the bowler  in different
venues
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get the data frame for RA Jadeja
jadeja <- getBowlerWicketDetails(team="India",name="Jadeja")
bowlerWicketsVenue(jadeja,"RA Jadeja")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{bowlerMovingAverage}}
\code{\link{bowlerWicketPlot}}
\code{\link{bowlerWicketsVenue}}
\code{\link{bowlerMeanRunsConceded}}
}

