% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlerWktsPredict.R
\name{bowlerWktsPredict}
\alias{bowlerWktsPredict}
\title{Compute the deliveries required for taking  the number of wickets}
\usage{
bowlerWktsPredict(df, name)
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of bowler}
}
\value{
None
}
\description{
This function uses a classification tree to compute the number deliveries needed
versus the wickets taken
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get the data frame for RA Jadeja
jadeja1 <- getDeliveryWickets(team="India",name="Jadeja",save=FALSE)
bowlerWktsPredict(jadeja1,"RA Jadeja")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{bowlerMovingAverage}}
\code{\link{bowlerWicketPlot}}
\code{\link{bowlerWicketsVenue}}
\code{\link{bowlerMeanRunsConceded}}
}

