% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertYaml2RDataframe.R
\name{convertYaml2RDataframe}
\alias{convertYaml2RDataframe}
\title{This function converts all Yaml files in source directory to data frames and saves as RData}
\usage{
convertYaml2RDataframe(yamlFile,sourceDir=".",targetDir=".")
}
\arguments{
\item{yamlFile}{The  yaml file to be converted to dataframe and saved}

\item{sourceDir}{The source directory of the yaml file}

\item{targetDir}{The target directory in which the data frame is stored as RData file}
}
\value{
None
}
\description{
This function coverts all Yaml files from source directory to data frames. The data frames
are then stored as .RData. The saved file is of the format team1-team2-date.RData
For e.g. England-India-2008-04-06.RData etc
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
# In the example below ../yamldir c
convertYaml2RDataframe("225171.yaml",".","../data")
}

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{bowlerMovingAverage}}
\code{\link{bowlerWicketPlot}}
\code{\link{bowlerWicketsVenue}}
\code{\link{convertAllYaml2RDataframes}}
}

