% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertYaml2RDataframeT20.R
\name{convertYaml2RDataframeT20}
\alias{convertYaml2RDataframeT20}
\title{Converts and save T20 yaml files to dataframes}
\usage{
convertYaml2RDataframeT20(yamlFile,sourceDir=".",targetDir=".")
}
\arguments{
\item{yamlFile}{The  yaml file to be converted to dataframe and saved}

\item{sourceDir}{The source directory of the yaml file}

\item{targetDir}{The target directory in which the data frame is stored as RData file}
}
\value{
None
}
\description{
This function coverts all T20 Yaml files from source directory to data frames. The data frames
are then stored as .RData. The saved file is of the format team1-team2-date.RData
For e.g. England-India-2008-04-06.RData etc
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
# In the example below ../yamldir c
convertYaml2RDataframe("225171.yaml",".","../data")
}

}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{bowlerMovingAverage}}\cr
\code{\link{bowlerWicketPlot}}\cr
\code{\link{bowlerWicketsVenue}}\cr
\code{\link{convertAllYaml2RDataframesT20}}\cr
}
\author{
Tinniam V Ganesh
}
