% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBatsmanDetails.R
\name{getBatsmanDetails}
\alias{getBatsmanDetails}
\title{Get the batting details of a batsman from the match}
\usage{
getBatsmanDetails(team,name)
}
\arguments{
\item{team}{The team of the batsman e.g. India}

\item{name}{Name of batsman}
}
\value{
None
}
\description{
This function gets the batting details of a batsman given the match data as a RData file
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{

getBatsmanDetails(team="India",name="Kohli")


}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{batsmanRunsPredict}}
\code{\link{batsmanMovingAverage}}
\code{\link{bowlerWicketsVenue}}
\code{\link{bowlerMeanRunsConceded}}
}

