% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCPLBattingDetails.R
\name{getCPLBattingDetails}
\alias{getCPLBattingDetails}
\title{Gets  the CPL batting details}
\usage{
getIPLBattingDetails(dir='.',odir=".")
}
\arguments{
\item{dir}{The input directory}

\item{odir}{The output directory}
}
\description{
This function creates a single datframe of all CPL batting
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{getBBLBattingDetails}}\cr
\code{\link{getBBLBowlingDetails}}\cr
\code{\link{getNTBBattingDetails}}\cr
\code{\link{getNTBBowlingDetails}}\cr
}
\author{
Tinniam V Ganesh
}
