% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatchDetails.R
\name{getMatchDetails}
\alias{getMatchDetails}
\title{Get  match details of 2 countries}
\usage{
getMatchDetails(team1,team2,date,dir=".")
}
\arguments{
\item{team1}{The 1st team in the match}

\item{team2}{The 2nd team in the matcj}

\item{date}{The date on which the match was played}

\item{dir}{The source directory of the RData files with all matches}
}
\value{
match
The dataframe of the match
}
\description{
This function  gets the details of a matc palyed between 2 countries from the saved RData files
and returns a dataframe
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
# convertAllYaml2RDataframes() & convertYaml2RDataframe convert yaml files
# to data frame and store as RData
# We have to point to this directory for the call below
a <- getMatchDetails("England","Pakistan","2006-09-05",dir="../data")

# Use this to create a classification tree of deliveries to wickets
bowlerWktsPredict(jadeja1,"RA Jadeja")
}

}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{getBatsmanDetails}}\cr
\code{\link{getBowlerWicketDetails}}\cr
\code{\link{getTeamBattingDetails}}\cr
\code{\link{getTeamBowlingDetails}}\cr
}
\author{
Tinniam V Ganesh
}
