% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNTBBowlingDetails.R
\name{getNTBBowlingDetails}
\alias{getNTBBowlingDetails}
\title{Gets  the NTB bowling details}
\usage{
getNTBBowlingDetails(dir='.',odir=".")
}
\arguments{
\item{dir}{The input directory}

\item{odir}{The output directory

\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}}
}
\description{
This function creates a single datframe of all NTB bowling
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\seealso{
\code{\link{getIPLBattingDetails}}\cr
\code{\link{getIPLBowlingDetails}}\cr
\code{\link{getNTBBattingDetails}}\cr
\code{\link{getNTBBattingDetails}}\cr
}
\author{
Tinniam V Ganesh
}
