% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWBBBowlingDetails.R
\name{getWBBBowlingDetails}
\alias{getWBBBowlingDetails}
\title{Gets  the WBB bowling details}
\usage{
getWBBBowlingDetails(dir='.',odir=".")
}
\arguments{
\item{dir}{The input directory}

\item{odir}{The output directory}
}
\description{
This function creates a single datframe of all WBB bowling
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{getIPLBattingDetails}}\cr
\code{\link{getIPLBowlingDetails}}\cr
\code{\link{getNTBBattingDetails}}\cr
\code{\link{getNTBBattingDetails}}\cr
}
\author{
Tinniam V Ganesh
}
