% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchWormChart.R
\name{matchWormGraph}
\alias{matchWormGraph}
\title{Plot the match worm graph}
\usage{
matchWormGraph(match,team1,team2)
}
\arguments{
\item{match}{The dataframe of the match}

\item{team1}{The 1st team of the match}

\item{team2}{the 2nd team in the match}
}
\value{
none
}
\description{
This function  plots the match worm graph between 2 teams in a match
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
#Get the match details
a <- getMatchDetails("England","Pakistan","2006-09-05",dir="../temp")

# Plot tne match worm plot
matchWormGraph(a,'England',"Pakistan")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{getBatsmanDetails}}\cr
\code{\link{getBowlerWicketDetails}}\cr
\code{\link{batsmanDismissals}}\cr
\code{\link{getTeamBattingDetails}}
}

