% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallRunsSRDeathOversPlotT20.R
\name{overallRunsSRDeathOversPlotT20}
\alias{overallRunsSRDeathOversPlotT20}
\title{Plot the Runs vs SR in  death overs  of Intl. T20 batsmen}
\usage{
overallRunsSRDeathOversPlotT20(dir=".", dateRange,type="IPL",plot=1)
}
\arguments{
\item{dir}{The input directory}

\item{dateRange}{Date interval to consider}

\item{type}{T20 league}

\item{plot}{plot=1 (static),plot=2(interactive), plot=3 (table)}

\item{odir}{The output directory}
}
\value{
The ranked T20 batsmen
}
\description{
Runs vs SR in  middle death  of Intl. T20 batsmen
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
overallRunsSRDeathOversPlotT20(dir=".", dateRange,type="IPL",plot=1)
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
