% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallRunsSRMiddleOversPlotT20M.R
\name{overallRunsSRMiddleOversPlotT20M}
\alias{overallRunsSRMiddleOversPlotT20M}
\title{Plot the Runs vs SR in  middle overs  of Intl. T20 batsmen}
\usage{
overallRunsSRMiddleOversPlotT20M(dir=".",minMatches, dateRange)
}
\arguments{
\item{minMatches}{Minimum matches played}

\item{dateRange}{Date interval to consider}

\item{teamNames}{The team names}

\item{odir}{The output directory}

\item{runsvsSR}{Runs or Strike rate}
}
\value{
The ranked T20 batsmen
}
\description{
Runs vs SR in  middle overs  of Intl. T20 batsmen
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
overallRunsSRMiddleOversPlotT20M(dir=".",dateRange)
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
