% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallWicketsERMiddleOversPlotT20M.R
\name{overallWicketsERMiddleOversPlotT20M}
\alias{overallWicketsERMiddleOversPlotT20M}
\title{Plot the Wickets vs ER in middle overs  of Intl. T20 batsmen}
\usage{
overallWicketsERMiddleOversPlotT20M(dir=".",minMatches, dateRange)
}
\arguments{
\item{minMatches}{Minimum matches played}

\item{dateRange}{Date interval to consider}

\item{teamNames}{The team names}

\item{odir}{The output directory}
}
\value{
The ranked T20 batsmen
}
\description{
Wickets vs ER in middle overs  of Intl. T20 batsmen
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
overallWicketsERPowerPlayPlotT20M(dir=".",minmatches, dateRange)
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
