% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallWicketsERPowerPlayPlotT20.R
\name{overallWicketsERPowerPlayPlotT20}
\alias{overallWicketsERPowerPlayPlotT20}
\title{Plot the Wickets vs ER in power play  of Intl. T20 batsmen}
\usage{
overallWicketsERPowerPlayPlotT20(dir=".", dateRange,type="IPL",plot=1)
}
\arguments{
\item{dir}{The input directory}

\item{dateRange}{Date interval to consider}

\item{type}{T20 league}

\item{plot}{plot=1 (static),plot=2(interactive), plot=3 (table)}
}
\description{
Wickets vs ER in power play  of Intl. T20 batsmen
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
overallWicketsERPowerPlayPlotT20(dir=".", dateRange,type="IPL",plot=1)
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
