% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseYamlOver.R
\name{parseYamlOver}
\alias{parseYamlOver}
\title{Parse yaml file and convert to dataframe}
\usage{
parseYamlOver(match,s,ateam,delivery,meta)
}
\arguments{
\item{match}{The dataframe of the match}

\item{s}{The string with the delivery}

\item{ateam}{The team}

\item{delivery}{The delivery of the over}

\item{meta}{The meta information of the match}
}
\value{
overs
The dataframe of overs
}
\description{
This function parses the yaml file and converts it into a data frame. This is an internal function and
is used by convertAllYaml2RDataframes() & convertYaml2RDataframe()
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
# Parse the yaml over
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{getBatsmanDetails}}\cr
\code{\link{getBowlerWicketDetails}}\cr
\code{\link{batsmanDismissals}}\cr
\code{\link{getTeamBattingDetails}}\cr
}
\author{
Tinniam V Ganesh
}
