% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBatsmenSeasonPerf.R
\name{plotBatsmenSeasonPerf}
\alias{plotBatsmenSeasonPerf}
\title{PLot the season performance of  T20 batsmen}
\usage{
plotBatsmenSeasonPerf(dir=".",batsmen, dateRange, type, runsvsSR,plot=1)
}
\arguments{
\item{dir}{The input directory}

\item{batsmen}{List of batsmen}

\item{dateRange}{Date interval to consider}

\item{type}{T20 format type}

\item{runsvsSR}{Runs or Strike rate}

\item{plot}{Whether to use ggplot2 or ggplotly}
}
\description{
This function creates a single datframe of all T20 batsmen and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
plotBatsmenSeasonPerf(dir=".",batsmen, dateRange,type, runsvsSR,plot=1)
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
