% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBowlersSeasonPerf.R
\name{plotBowlersSeasonPerf}
\alias{plotBowlersSeasonPerf}
\title{PLot the season performance of  T20 batsmen}
\usage{
plotBowlersSeasonPerf(dir=".",bowlers, dateRange, type, wicketsVsER,plot=1)
}
\arguments{
\item{dir}{The input directory}

\item{bowlers}{List of bowlers}

\item{dateRange}{Date interval to consider}

\item{type}{T20 format type}

\item{wicketsVsER}{Wickets or Economy Rate}

\item{plot}{Whether to use ggplot2 or ggplotly}
}
\description{
This function creates a single datframe of all T20 batsmen and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
plotBowlersSeasonPerf(dir=".",bowlers, dateRange, type, wicketsVsER,plot=1)
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
