% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWinLossBetweenTeams.R
\name{plotWinLossBetweenTeams}
\alias{plotWinLossBetweenTeams}
\title{This function computes and plots number of wins for each team in all their
encounters}
\usage{
plotWinLossBetweenTeams(team1,team2,dir=".")
}
\arguments{
\item{team1}{The 1st team}

\item{team2}{The 2nd team}

\item{dir}{The source directory of teh RData files}
}
\value{
None
}
\description{
This function computes and plots number of wins for each team in all their
encounters. The plot includes the number of  wins by each team and the matches
with no result
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{

plotWinLossBetweenTeams(team1="India",team2="Australia",dir=pathToFile)
batsmanDismissals(kohli,"Kohli")
}
}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{batsmanFoursSixes}}\cr
\code{\link{batsmanRunsVsDeliveries}}\cr
\code{\link{batsmanRunsVsStrikeRate}}\cr
}

