% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWinLossTeamVsAllTeams.R
\name{plotWinLossTeamVsAllTeams}
\alias{plotWinLossTeamVsAllTeams}
\title{Plot  wins for each team}
\usage{
plotWinLossTeamVsAllTeams(team1,dir=".")
}
\arguments{
\item{team1}{The 1st team}

\item{dir}{The source directory of the RData files}
}
\value{
None
}
\description{
This function computes and plots number of wins for a team against all
other teams. The plot includes the number of  wins by team each team and the matches
with no result
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{

plotWinLossBetweenTeams(team1="India",team2="Australia",dir=pathToFile)
batsmanDismissals(kohli,"Kohli")
}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{batsmanFoursSixes}}\cr
\code{\link{batsmanRunsVsDeliveries}}\cr
\code{\link{batsmanRunsVsStrikeRate}}\cr
}
\author{
Tinniam V Ganesh
}
