% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankBBLBowlers.R
\name{rankBBLBowlers}
\alias{rankBBLBowlers}
\title{Ranks the BBL bowlers}
\usage{
rankBBLBowlers(dir='.',odir=".",minMatches=20)
}
\arguments{
\item{dir}{The input directory}

\item{odir}{The output directory}

\item{minMatches}{Minimum matches}
}
\value{
The ranked BBL bowlers
}
\description{
This function creates a single datframe of all BBL bowlers and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#
bblBowlersRank <- rankBBLBowlers()
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{rankIPLBatsmen}}\cr
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Batsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
