% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankIPLBowlers.R
\name{rankIPLBowlers}
\alias{rankIPLBowlers}
\title{Ranks the IPL bowlers}
\usage{
rankIPLBowlers()
}
\value{
The ranked IPL batsmen
}
\description{
This function creates a single datframe of all IPL bowlers and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#
iplBowlersRank <- rankIPLBowlers()
}

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{rankIPLBatsmen}}\cr
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Batsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}

