% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankIPLBowlers1.R
\name{rankIPLBowlers1}
\alias{rankIPLBowlers1}
\title{Ranks the IPL bowlers1}
\usage{
rankIPLBowlers1(dir='.',odir=".",minMatches=20)
}
\arguments{
\item{dir}{The input directory}

\item{odir}{The output directory}

\item{minMatches}{Minimum matches}
}
\value{
The ranked IPL batsmen
}
\description{
This function creates a single datframe of all IPL bowlers and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#
iplBowlersRank <- rankIPLBowlers()
}

}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{rankIPLBatsmen}}\cr
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Batsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
