% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankNTBBatsmen.R
\name{rankNTBBatsmen}
\alias{rankNTBBatsmen}
\title{Ranks the NTB batsmen}
\usage{
rankNTBBatsmen(dir='.',odir=".",minMatches=50)
}
\arguments{
\item{dir}{The input directory}

\item{odir}{The output directory}

\item{minMatches}{Minimum matches}
}
\value{
The ranked NTB batsmen
}
\description{
This function creates a single datframe of all NTB batsmen and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#
ntbBatsmanRank <- rankNTBBatsmen()
}

}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{rankIPLBowlers}}\cr
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Batsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
