% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankODIBatsmen.R
\name{rankODIBatsmen}
\alias{rankODIBatsmen}
\title{Ranks the ODI batsmen}
\usage{
rankODIBatsmen(dir='.',odir=".",minMatches=50)
}
\arguments{
\item{dir}{The input directory}

\item{odir}{The output directory}

\item{minMatches}{Minimum matches}
}
\value{
The ranked ODI batsmen
}
\description{
This function creates a single datframe of all ODI batsmen and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
odiBatsmanRank <- rankODIBatsmen()
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankODIBowlers}}\cr
\code{\link{rankT20Batsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
