% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankODIBowlers.R
\name{rankODIBowlers}
\alias{rankODIBowlers}
\title{Ranks the ODI bowlers}
\usage{
rankODIBowlers(dir=".",odir=".")
}
\arguments{
\item{dir}{The input directory}

\item{odir}{The output directory}
}
\value{
The ranked ODI bowlers
}
\description{
This function creates a single datframe of all ODI bowlers and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#
odiBowlersRank <- rankODIBowlers()
}

}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{rankIPLBowlers}}\cr
\code{\link{rankIPLBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Batsmen}}\cr
}
\author{
Tinniam V Ganesh
}
