% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankPSLBowlers.R
\name{rankPSLBowlers}
\alias{rankPSLBowlers}
\title{Ranks the PSL bowlers}
\usage{
rankPSLBowlers(dir='.',odir=".",minMatches=20)
}
\arguments{
\item{dir}{The input directory}

\item{odir}{The output directory}

\item{minMatches}{Minimum matches}
}
\value{
The ranked PSL batsmen
}
\description{
This function creates a single datframe of all PSL bowlers and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#
pslBowlersRank <- rankPSLBowlers()
}

}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{rankIPLBatsmen}}\cr
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Batsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
