% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankT20Batsmen.R
\name{rankT20Batsmen}
\alias{rankT20Batsmen}
\title{Ranks the T20 batsmen}
\usage{
rankT20Batsmen(dir='.',odir=".",minMatches=50)
}
\arguments{
\item{odir}{The output directory}

\item{minMatches}{Minimum matches}

\item{dir}{The input directory}
}
\value{
The ranked T20 batsmen
}
\description{
This function creates a single datframe of all T20 batsmen and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#
t20BatsmanRank <- rankT20Batsmen()
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankIPLBowlers}}\cr
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
