% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankWBBBowlers.R
\name{rankWBBBowlers}
\alias{rankWBBBowlers}
\title{Ranks the WBB bowlers}
\usage{
rankWBBBowlers(dir='.',odir=".",minMatches=20)
}
\arguments{
\item{dir}{The input directory}

\item{odir}{The output directory}

\item{minMatches}{Minimum matches}
}
\value{
The ranked WBB batsmen
}
\description{
This function creates a single datframe of all WBB bowlers and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#
WBBBowlersRank <- rankWBBBowlers()
}

}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{rankIPLBatsmen}}\cr
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Batsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
