% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamRunsSRMiddleOversPlotAllOppnAllMatches.R
\name{runsSRPMiddleOversPlotAllOppnAllMatches}
\alias{runsSRPMiddleOversPlotAllOppnAllMatches}
\title{Team batting plots runs vs SR in middle overs for team against all oppositions in all matches}
\usage{
teamRunsSRMiddleOversPlotAllOppnAllMatches(matches,theTeam)
}
\arguments{
\item{t1}{The team for which the the batting partnerships are sought}

\item{matches}{All matches of the team in all matches with all oppositions}
}
\value{
details
The data frame of the scorecard of the team in all matches against all oppositions
}
\description{
This function computes and plots runs vs SR  in middle overs of a team in all matches against all
oppositions.
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{


# Top batsman is displayed in descending order of runs
teamRunsSRMiddleOversPlotAllOppnAllMatches(matches,t1="India")

}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{teamBatsmenVsBowlersAllOppnAllMatchesPlot}}\cr
\code{\link{teamBatsmenPartnershipOppnAllMatchesChart}}\cr
\code{\link{teamBatsmenPartnershipAllOppnAllMatchesPlot}}\cr
\code{\link{teamBowlingWicketRunsAllOppnAllMatches}}
}
\author{
Tinniam V Ganesh
}
