% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAllIPLBattingDetails.R
\name{saveAllIPLBattingDetails}
\alias{saveAllIPLBattingDetails}
\title{Save all IPL batting details}
\usage{
saveAllIPLBattingDetails(teamNames,dir=".",odir=".",type="IPL",save=TRUE)
}
\arguments{
\item{teamNames}{The team names}

\item{dir}{The output directory}

\item{odir}{The output directory}
}
\value{
The ranked T20 batsmen
}
\description{
This function creates a single dataframe of all IPL batting details
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
saveAllIPLBattingDetails(teamNames,dir=".",odir=".",type="IPL",save=TRUE)
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
