% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAllMatchesAllOpposition.R
\name{saveAllMatchesAllOpposition}
\alias{saveAllMatchesAllOpposition}
\title{This function saves all matches against all opposition to a team as a
single data frame}
\usage{
saveAllMatchesAllOpposition()
}
\value{
None
}
\description{
This function saves all matches agaist all opposition as a single dataframe in the
current directory
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
saveAllMatchesBetweenTeams
}
}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{batsmanDismissals}}\cr
\code{\link{batsmanRunsVsDeliveries}}\cr
\code{\link{batsmanRunsVsStrikeRate}}\cr
\code{\link{getAllMatchesAllOpposition}}\cr
\code{\link{getAllMatchesBetweenTeams}}\cr
}

