% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAllMatchesAllOppositionIPLT20.R
\name{saveAllMatchesAllOppositionIPLT20}
\alias{saveAllMatchesAllOppositionIPLT20}
\title{Saves matches against all IPL teams as dataframe for an IPL team}
\usage{
saveAllMatchesAllOppositionIPLT20()
}
\value{
None
}
\description{
This function saves all IPL matches agaist all opposition as a single dataframe in the
current directory
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
saveAllMatchesAllOppositionT20
}
}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{batsmanDismissals}}\cr
\code{\link{batsmanRunsVsDeliveries}}\cr
\code{\link{batsmanRunsVsStrikeRate}}\cr
\code{\link{getAllMatchesAllOpposition}}\cr
\code{\link{getAllMatchesBetweenTeams}}\cr
}

