% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAllMatchesAllOppositionWBBT20.R
\name{saveAllMatchesAllOppositionWBBT20}
\alias{saveAllMatchesAllOppositionWBBT20}
\title{Saves matches against all WBB teams as dataframe for an WBB team}
\usage{
saveAllMatchesAllOppositionWBBT20(dir=".",odir=".")
}
\arguments{
\item{dir}{Input Directory}

\item{odir}{Output Directory}
}
\value{
None
}
\description{
This function saves all WBB matches agaist all opposition as a single dataframe in the
output directory
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
saveAllMatchesAllOppositionT20
}
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{batsmanDismissals}}\cr
\code{\link{batsmanRunsVsDeliveries}}\cr
\code{\link{batsmanRunsVsStrikeRate}}\cr
\code{\link{getAllMatchesAllOpposition}}\cr
\code{\link{getAllMatchesBetweenTeams}}\cr
}
\author{
Tinniam V Ganesh
}
