% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAllMatchesBetween2BBLTeams.R
\name{saveAllMatchesBetween2BBLTeams}
\alias{saveAllMatchesBetween2BBLTeams}
\title{Saves all matches between 2 BBL teams as  dataframe}
\usage{
saveAllMatchesBetween2BBLTeams(dir=".",odir=".")
}
\arguments{
\item{dir}{Input Directory}

\item{odir}{Output Directory to store saved matches}
}
\value{
None
}
\description{
This function saves all matches between 2 BBL teams as a single dataframe in the
current directory
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
saveAllMatchesBetween2BBLTeams(dir=".",odir=".")
}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{batsmanDismissals}}\cr
\code{\link{batsmanRunsVsDeliveries}}\cr
\code{\link{batsmanRunsVsStrikeRate}}\cr
\code{\link{getAllMatchesAllOpposition}}\cr
\code{\link{getAllMatchesBetweenTeams}}\cr
}
\author{
Tinniam V Ganesh
}
