% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAllMatchesBetweenTeams.R
\name{saveAllMatchesBetweenTeams}
\alias{saveAllMatchesBetweenTeams}
\title{Saves all matches between 2 teams as  dataframe}
\usage{
saveAllMatchesBetweenTeams()
}
\value{
None
}
\description{
This function saves all matches between 2 teams as a single dataframe in the
current directory
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
saveAllMatchesBetweenTeams
}
}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{batsmanDismissals}}\cr
\code{\link{batsmanRunsVsDeliveries}}\cr
\code{\link{batsmanRunsVsStrikeRate}}\cr
\code{\link{getAllMatchesAllOpposition}}\cr
\code{\link{getAllMatchesBetweenTeams}}\cr
}

