% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAllT20BowlingDetails.R
\name{saveAllT20BowlingDetails}
\alias{saveAllT20BowlingDetails}
\title{Save all T20 batting details}
\usage{
saveAllT20BowlingDetails(teamNames,dir=".",odir=".",type="IPL",save=TRUE)
}
\arguments{
\item{teamNames}{The team names}

\item{dir}{The output directory}

\item{odir}{The output directory}

\item{save}{To save or not}
}
\description{
This function creates a single dataframe of all T20 batting details
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
saveAllT20BowlingDetails(teamNames,dir=".",odir=".",type="IPL",save=TRUE)
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
