% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAllT20MatchesAsDF.R
\name{saveAllT20MatchesAsDF}
\alias{saveAllT20MatchesAsDF}
\title{Overall picture Ranks the T20 batsmen}
\usage{
saveAllT20MatchesAsDF(teamNames,dir=".",odir=".",type="IPL",save=TRUE)
}
\arguments{
\item{teamNames}{The team names}

\item{dir}{The output directory}

\item{odir}{The output directory}

\item{save}{To save or not}
}
\description{
This function creates a single datframe of all T20 batsmen and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
saveAllT20MatchesAsDF(teamNames,dir=".",odir=".",type="IPL",save=TRUE)
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
