% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialProc.R
\name{specialProc}
\alias{specialProc}
\title{Used to parse yaml file}
\usage{
specialProc(dist, overset, ateam,over,str1,meta)
}
\arguments{
\item{dist}{dist}

\item{overset}{overset}

\item{ateam}{The team}

\item{over}{over}

\item{str1}{str1}

\item{meta}{The meta information of the match}
}
\value{
over
The dataframe of over
}
\description{
This is special processing function. This is an internal function and
is used by convertAllYaml2RDataframes() & convertYaml2RDataframe()
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
# Parse the yaml over
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{getBatsmanDetails}}\cr
\code{\link{getBowlerWicketDetails}}\cr
\code{\link{batsmanDismissals}}\cr
\code{\link{getTeamBattingDetails}}\cr
}
\author{
Tinniam V Ganesh
}
