% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBatsmanPartnershipAllOppnAllMatchesPlot.R
\name{teamBatsmanPartnershipAllOppnAllMatchesPlot}
\alias{teamBatsmanPartnershipAllOppnAllMatchesPlot}
\title{This function plots the batting partnership of a team in all matches against all
oppositions or it returns the data frame}
\usage{
teamBatsmanPartnershipAllOppnAllMatchesPlot(matches,theTeam,main,plot=TRUE)
}
\arguments{
\item{matches}{All the matches of the team against all oppositions}

\item{theTeam}{The team for which the the batting partnerships are sought}

\item{plot}{Whether the partnerships have top be rendered as a plot. If plot=FALSE the data frame is returned}
}
\value{
None or partnerships
}
\description{
This function plots  the batting partnership of a team againt all oppositions in all matches
This function also returns a  dataframe  with the batting partnerships
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get all matches for team India against all oppositions
d <- teamBatsmanVsBowlersAllOppnAllMatchesRept(matches,"India",rank=1,dispRows=50)
 #Plot the partnerships
teamBatsmanVsBowlersAllOppnAllMatchesPlot(d)

#Do not plot but get the dataframe
e <- teamBatsmanVsBowlersAllOppnAllMatchesPlot(d,plot=FALSE)

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{teamBatsmanPartnershipAllOppnAllMatches}}
\code{\link{teamBatsmanPartnershipOppnAllMatchesChart}}
\code{\link{teamBatsmanvsBowlersAllOppnAllMatchesPlot}}
\code{\link{teamBatsmanVsBowlerOppnAllMatches}}
}

