% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBatsmenPartnershipOppnAllMatches.R
\name{teamBatsmenPartnershiOppnAllMatches}
\alias{teamBatsmenPartnershiOppnAllMatches}
\title{This function computes  and returns the partenership of a team of  batsmen against bowlers in
all matches against an opposition}
\usage{
teamBatsmenPartnershiOppnAllMatches(matches,theTeam,report="summary")
}
\arguments{
\item{matches}{All the matches of the team against the oppositions}

\item{theTeam}{The team for which the the batting partnerships are sought}

\item{report}{If the report="summary" then the list of top batsmen with the highest partnerships is displayed. If
report="detailed" then the detailed break up of partnership is returned as a dataframe}
}
\value{
partnerships
The data frame of the partnerships
}
\description{
This function computes the performance of batsmen against all bowlers of an oppositions in all matches. This
function returns a dataframe
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get all matches for team India against all oppositions
matches <- getAllMatchesBetweenTeams("Australia","India",dir="../data")
You can also directly load the data
#load("India-Australia-allMatches.RData")

m <-teamBatsmenPartnershiOppnAllMatches(a,'India',report="summary")
m <-teamBatsmenPartnershiOppnAllMatches(a,'Australia',report="detailed")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{teamBatsmanvsBowlersAllOppnAllMatchesPlot}}
\code{\link{teamBatsmanPartnershipOppnAllMatchesChart}}
\code{\link{teamBowlerVsBatsmanMatch}}
\code{\link{teamBattingScorecardMatch}}
}

