% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBatsmenVsBowlersMatch.R
\name{teamBatsmenVsBowlersMatch}
\alias{teamBatsmenVsBowlersMatch}
\title{This function plots the performance of batsmen against bowlers in a match . It can also return
the data frame}
\usage{
teamBatsmenVsBowlersMatch(match,theTeam,plot=TRUE)
}
\arguments{
\item{match}{The match between the teams}

\item{theTeam}{The team for which the the batting partnerships are sought}

\item{plot}{If plot=TRUE then a plot is created otherwise a data frame is returned}
}
\value{
b
The data frame of the batsmen vs bowlers performance
}
\description{
This function plots the performance of  batsmen versus bowlers in a match  or it can return
the data frame
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
# Get athe match between England and Pakistan
a <- getMatchDetails("England","Pakistan","2006-09-05",dir="../temp")
batsmenVsBowlersMatch(a,'Pakistan',plot=TRUE)
teamBowlingScorecardMatch(a,'England')
teamBowlingWicketKindMatch(a,"England",plot=FALSE)
}

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{teamBatsmenVsBowlersAllOppnAllMatchesPlot}}\cr
\code{\link{teamBatsmenPartnershipOppnAllMatchesChart}}\cr
\code{\link{teamBatsmenPartnershipAllOppnAllMatchesPlot}}\cr
\code{\link{teamBattingScorecardMatch}}\cr
}

