% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBattingScoredcardOppnAllMatches.R
\name{teamBattingScorecardOppnAllMatches}
\alias{teamBattingScorecardOppnAllMatches}
\title{Compute and return the batting scorecard of a team in all matches against an opposition}
\usage{
teamBattingScorecardMatch(matches,main,opposition)
}
\arguments{
\item{matches}{the data frame of all matches between a team and an opposition obtained with
the call getAllMatchesBetweenteam()}

\item{main}{The main team for which scorecard required}

\item{opposition}{The opposition team}
}
\value{
scorecard
The scorecard of all the matches
}
\description{
This function computes returns the batting scorecard (runs, fours, sixes, balls played) for the
team in all matches against an opposition
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get all matches between India and Australia
matches <- getAllMatchesBetweenTeams("India","Australia",dir="../data",save=TRUE)
# Compute the scorecard of India in matches with australia
teamBattingScorecardOppnAllMatches(matches,main="India",opposition="Australia")

#Get all matches between Australia and India
matches <- getAllMatchesBetweenTeams("Australia","India",dir="../data")
#Compute the batting scorecard of Australia
teamBattingScorecardOppnAllMatches(matches,"Australia","India")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{teamBatsmanPartnershipAllOppnAllMatches}}
\code{\link{teamBatsmanPartnershipAllOppnAllMatchesPlot}}
\code{\link{teamBatsmanPartnershipOppnAllMatchesChart}}
}

