% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBowlerVsBatsmanOppnAllMatches.R
\name{teamBowlersVsBatsmanOppnAllMatches}
\alias{teamBowlersVsBatsmanOppnAllMatches}
\title{Compute the performance of bowlers of a team against an opposition in all matches}
\usage{
teamBowlersVsBatsmanOppnAllMatches(match,main,opposition,plot=TRUE,top=5)
}
\arguments{
\item{match}{The data frame of all matches between a team the opposition. This dataframe can be obtained with
matches <- getAllMatchesBetweenTeams("Australia","India",dir="../data")}

\item{theTeam}{The team against which the performance is requires}

\item{rank}{When the rank is 0 then the performance of all the bowlers is displayed. If rank=n (1,2,3 ..) then
the performance of that bowler is given}
}
\value{
dataframe
The dataframe with all performances
}
\description{
This function computes performance of bowlers of a team against an opposition in all matches
against the opposition
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get all matches between India and Australia
matches <- getAllMatchesBetweenTeams("Australia","India",dir="../data")

#  Plot the performance of top 5 Indian bowlers against Australia
teamBowlersVsBatsmanOppnAllMatches(matches,'India',"Australia",top=5)

# Plot the performance of top 3 Australian bowlers against India
teamBowlersVsBatsmanOppnAllMatches(matches,"Australia","India",top=3)

# Get the top 5 bowlers of Australia. Do not plot but get as a dataframe
teamBowlersVsBatsmanOppnAllMatches(matches,"Australia","India",plot=FALSE)

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{teamBatsmanPartnershipAllOppnAllMatches}}
\code{\link{teamBatsmanPartnershipAllOppnAllMatchesPlot}}
\code{\link{teamBatsmanPartnershipOppnAllMatchesChart}}
\code{\link{teamBowlersVsBatsmanAllOppnAllMatchesRept}}
\code{\link{teamBowlersVsBatsmanAllOppnAllMatchesPlot}}
}

