% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBowlersVsBatsmenAllOppnAllMatchesMain.R
\name{teamBowlersVsBatsmenAllOppnAllMatchesMain}
\alias{teamBowlersVsBatsmenAllOppnAllMatchesMain}
\title{Compute team bowlers vs batsmen all opposition  all matches}
\usage{
teamBowlersVsBatsmenAllOppnAllMatchesMain(matches,theTeam,rank=0)
}
\arguments{
\item{matches}{the data frame of all matches between a team and aall opposition  and all obtained with
the call getAllMatchesAllOpposition()}

\item{theTeam}{The team against which the performance is requires}

\item{rank}{When the rank is 0 then the performance of all the bowlers is displayed. If rank=n (1,2,3 ..) then
the performance of that bowler is given}
}
\value{
dataframe
The dataframe with all performances
}
\description{
This function computes performance of bowlers of a team against all opposition in all matches
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
# Get all matches between India and all oppostions
matches <-getAllMatchesAllOpposition("India",dir="../data/",save=TRUE)

# You could also load directly from the saved file
#load("allMatchesAllOpposition-India.RData")
# The call below gives the best bowlers of India
teamBowlersVsBatsmenAllOppnAllMatchesMain(matches,theTeam="India",rank=0)

# The call with rank=1 gives the performance of the 'India' bowler with rank=1
 teamBowlersVsBatsmenAllOppnAllMatchesMain(matches,theTeam="India",rank=1)
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{teamBatsmenPartnershipAllOppnAllMatches}}\cr
\code{\link{teamBatsmenPartnershipAllOppnAllMatchesPlot}}\cr
\code{\link{teamBatsmenPartnershipOppnAllMatchesChart}}\cr
\code{\link{teamBowlersVsBatsmenAllOppnAllMatchesRept}}\cr
\code{\link{teamBowlersVsBatsmenAllOppnAllMatchesPlot}}\cr
}
\author{
Tinniam V Ganesh
}
