% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBowlersVsBatsmenMatch.R
\name{teamBowlersVsBatsmenMatch}
\alias{teamBowlersVsBatsmenMatch}
\title{Compute the performance of bowlers of a team in a specific match}
\usage{
teamBowlersVsBatsmenMatch(match,theTeam,plot=TRUE)
}
\arguments{
\item{match}{The data frame of the match. This can be obtained with the call for e.g
a <- getMatchDetails("England","Pakistan","2006-09-05",dir="../temp")}

\item{theTeam}{The team against which the performance is required}

\item{plot}{This parameter specifies if a plot is required, If plot=FALSE then a data frame is returned}
}
\value{
None or dataframe
If plot=TRUE there is no return. If plot=TRUE then the dataframe is returned
}
\description{
This function computes performance of bowlers of a team against an opposition in a match
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get the match between England and Pakistan
a <- getMatchDetails("England","Pakistan","2006-09-05",dir="../temp")

teamBowlersVsBatsmenMatch(a,"Pakistan")
teamBowlersVsBatsmenMatch(a,"England")
m <- teamBowlersVsBatsmenMatch(a,"Pakistan")



}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{teamBatsmenPartnershipAllOppnAllMatches}}
\code{\link{teamBatsmenPartnershipAllOppnAllMatchesPlot}}
\code{\link{teamBatsmenPartnershipOppnAllMatchesChart}}
\code{\link{teamBowlersVsBatsmenAllOppnAllMatchesRept}}
\code{\link{teamBowlersVsBatsmenAllOppnAllMatchesPlot}}
}

