% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBowlerVsBatsmanAllOppnAllMatchesPlot.R
\name{teamBowlerVsBatsmanAllOppnAllMatchesPlot}
\alias{teamBowlerVsBatsmanAllOppnAllMatchesPlot}
\title{Compute the performance of bowlers of a team against all opposition in all matches}
\usage{
teamBowlerVsBatsmanAllOppnAllMatchesPlot(bowlerDF,t1,t2)
}
\arguments{
\item{t1}{The team against to which the player belong}

\item{t2}{The opposing team}

\item{matches}{the data frame of the bowler whose performance is required}
}
\value{
none
}
\description{
This function computes performance of bowlers of a team against all opposition in all matches
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get all matches between India and all oppostions
matches <-getAllMatchesAllOpposition("India",dir="../data/",save=TRUE)

#Get the details of the bowler with the specified rank as a dataframe
df <- teamBowlersVsBatsmanAllOppnAllMatchesRept(matches,theTeam="India",rank=1)
#Plot this
teamBowlerVsBatsmanAllOppnAllMatchesPlot(df,"India","India")

df <- teamBowlersVsBatsmanAllOppnAllMatchesRept(matches,theTeam="England",rank=1)
teamBowlerVsBatsmanAllOppnAllMatchesPlot(df,"India","England")


}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{teamBatsmanPartnershipAllOppnAllMatches}}
\code{\link{teamBatsmanPartnershipAllOppnAllMatchesPlot}}
\code{\link{teamBatsmanPartnershipOppnAllMatchesChart}}
\code{\link{teamBowlersVsBatsmanAllOppnAllMatchesMain}}
\code{\link{teamBowlersVsBatsmanAllOppnAllMatchesRept}}
}

