% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBowlerWicketKindOppnAllMatches.R
\name{teamBowlerWicketKindOppnAllMatches}
\alias{teamBowlerWicketKindOppnAllMatches}
\title{Compute the performance of bowlers of a team  and the wicket kind against an
opposition in all matches}
\usage{
teamBowlerWicketKindOppnAllMatches(matches,main,opposition,plot=TRUE)
}
\arguments{
\item{matches}{The data frame of all matches between a team the opposition. This dataframe can be obtained with
matches <- getAllMatchesBetweenTeams("Australia","India",dir="../data")}

\item{main}{The team for which the performance is required}

\item{opposition}{The opposing team}

\item{plot}{If plot=TRUE then a plot is displayed else a dataframe is returned}
}
\value{
None or dataframe
The return depends on the value of the plot
}
\description{
This function computes performance of bowlers of a team and the wicket kind against an
opposition in all matches against the opposition
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get all matches between India and Australia
matches <- getAllMatchesBetweenTeams("Australia","India",dir="../data")

teamBowlerWicketKindOppnAllMatches(matches,"India","Australia",plot=TRUE)
m <- teamBowlerWicketKindOppnAllMatches(matches,"Australia","India",plot=FALSE)

teamBowlerWicketKindOppnAllMatches(matches,"Australia","India",plot=TRUE)

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{teamBatsmanPartnershipAllOppnAllMatches}}
\code{\link{teamBatsmanPartnershipAllOppnAllMatchesPlot}}
\code{\link{teamBatsmanPartnershipOppnAllMatchesChart}}
\code{\link{teamBowlersVsBatsmanAllOppnAllMatchesRept}}
\code{\link{teamBowlersVsBatsmanAllOppnAllMatchesPlot}}
}

