% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBowlerWicketsOppnAllMatches.R, R/teamBowlingWicketsOppnAllMatches.R
\name{teamBowlerWicketsOppnAllMatches}
\alias{teamBowlerWicketsOppnAllMatches}
\title{Compute the performance of bowlers of a team  and the wicket taken against an
opposition in all matches}
\usage{
teamBowlerWicketsOppnAllMatches(matches,main,opposition,plot=TRUE,top=20)

teamBowlerWicketsOppnAllMatches(matches,main,opposition,plot=TRUE,top=20)
}
\arguments{
\item{matches}{The data frame of all matches between a team the opposition. This dataframe can be obtained with
matches <- getAllMatchesBetweenTeams("Australia","India",dir="../data")}

\item{main}{The team for which the performance is required}

\item{opposition}{The opposing team}

\item{plot}{If plot=TRUE then a plot is displayed else a dataframe is returned}

\item{top}{The number of top bowlers to be included in the result}

\item{matches}{The matches between the teams}

\item{main}{Team for which bowling performance is required}

\item{opposition}{The oppositionteam}

\item{plot}{If plot= TRUE the dataframe will be plotted else a data frame will be returned

top
The number of rows of data frame to be returned, Default is 20}
}
\value{
None or dataframe
The return depends on the value of the plot

None or data fame
}
\description{
This function computes performance of bowlers of a team and the wickets taken against an
opposition in all matches against the opposition

This function computes returns the wickets taken against an opposition in all matches
The user can choose to plot or return a dataframe
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}

Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get all matches between India and Australia
matches <- getAllMatchesBetweenTeams("Australia","India",dir="../data")

#Display top 20
teamBowlerWicketsOppnAllMatches(matches,"India","Australia")
#Display and plot top 10
teamBowlerWicketsOppnAllMatches(matches,"Australia","India",top=10)

#Do not plot but return as dataframe
teamBowlerWicketsOppnAllMatches(matches,"India","Australia",plot=FALSE)


#Get the match details
matches <- getAllMatchesBetweenTeams("Australia","India",dir="../data")

# Or the data can be directly loaded
#load("India-Australia-allMatches.RData")

teamBowlerWicketsOppnAllMatches(matches,"India","Australia")
teamBowlerWicketsOppnAllMatches(matches,"Australia","India",top=10)
teamBowlerWicketsOppnAllMatches(matches,"India","Australia",plot=FALSE)

}
\author{
Tinniam V Ganesh

Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}

\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{teamBatsmanPartnershipAllOppnAllMatches}}
\code{\link{teamBatsmanPartnershipAllOppnAllMatchesPlot}}
\code{\link{teamBatsmanPartnershipOppnAllMatchesChart}}
\code{\link{teamBowlersVsBatsmanAllOppnAllMatchesRept}}
\code{\link{teamBowlerWicketRunsOppnAllMatches}}

\code{\link{teamBowlingWicketMatch}}\cr
\code{\link{teamBowlingWicketRunsMatch}}\cr
\code{\link{teamBowlerVsBatsmanMatch}}\cr
}

