% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBowlingPerfDetails.R
\name{teamBowlingPerfDetails}
\alias{teamBowlingPerfDetails}
\title{get team bowling performance details}
\usage{
teamBowlingPerfDetails(match,theTeam,includeInfo=FALSE)
}
\arguments{
\item{match}{The data frame of all match}

\item{theTeam}{The team for which the performance is required}

\item{includeInfo}{If true details like venie,winner, result etc are included}
}
\value{
dataframe
The dataframe of bowling performance
}
\description{
This function computes performance of bowlers of a team a
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
# Get all matches between India and Australia
match <- getMatchDetails("England","Pakistan","2006-09-05",dir="../temp")
teamBowlingPerf(match,"India",includeInfo=TRUE)
}

}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{teamBatsmenPartnershipAllOppnAllMatches}}\cr
\code{\link{teamBatsmenPartnershipAllOppnAllMatchesPlot}}\cr
\code{\link{teamBatsmenPartnershipOppnAllMatchesChart}}\cr
\code{\link{teamBowlersVsBatsmenAllOppnAllMatchesRept}}\cr
\code{\link{teamBowlersWicketRunsOppnAllMatches}}\cr
}
\author{
Tinniam V Ganesh
}
