% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBowlingScorecardAllOppnAllMatches.R
\name{teamBowlingScorecardAllOppnAllMatches}
\alias{teamBowlingScorecardAllOppnAllMatches}
\title{Compute and return a data frame of the bowling scorecard of a  team against all oppositions in all matches}
\usage{
teamBowlingScorecardAllOppnAllMatches(matches,theTeam)
}
\arguments{
\item{matches}{The matches of the team against all oppositions and all matches}

\item{theTeam}{Team for which bowling performance is required}
}
\value{
l
A data frame with the bowling performance in alll matches against all oppositions
}
\description{
This function computes returns the bowling dataframe of bowlers deliveries, maidens, overs, wickets
against all oppositions in all matches
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#Get all matches between India  and other opposition
matches <-getAllMatchesAllOpposition("India",dir="../data/",save=TRUE)

# Or load directly from saved file
# load("allMatchesAllOpposition-India.RData")

# Top opposition bowlers performances against India
teamBowlingScorecardAllOppnAllMatches(matches,"India")

#Top Indian bowlers against respective opposition
teamBowlingScorecardAllOppnAllMatches(matches,'Australia')
teamBowlingScorecardAllOppnAllMatches(matches,'South Africa')
teamBowlingScorecardAllOppnAllMatches(matches,'England')
}

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{teamBowlingScorecardAllOppnAllMatchesMain}}\cr
\code{\link{teamBowlersVsBatsmenAllOppnAllMatchesMain}}\cr
\code{\link{teamBowlersVsBatsmenAllOppnAllMatchesPlot}}\cr
}

