% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBowlingScorecardAllOppnAllMatchesMain.R
\name{teamBowlingScorecardAllOppnAllMatchesMain}
\alias{teamBowlingScorecardAllOppnAllMatchesMain}
\title{Compute and return a data frame of the best bowlers of a  team against all oppositions in all matches}
\usage{
teamBowlingScorecardAllOppnAllMatchesMain(matches,theTeam)
}
\arguments{
\item{matches}{The matches of the team against all oppositions and all matches}

\item{theTeam}{Team for which bowling performance is required}
}
\value{
l
A data frame with the bowling performance in alll matches against all oppositions
}
\description{
This function computes returns the bowling dataframe of best bowlers deliveries, maidens, overs, wickets
against all oppositions in all matches
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
#Get all matches between India  and other opposition
matches <-getAllMatchesAllOpposition("India",dir="../data/",save=TRUE)

# Or load directly from saved file
# load("allMatchesAllOpposition-India.RData")

# Top opposition bowlers of India
teamBowlingScorecardAllOppnAllMatchesMain(matches,"India")


}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{teamBowlingScorecardAllOppnAllMatches}}\cr
\code{\link{teamBowlersVsBatsmenAllOppnAllMatchesMain}}\cr
\code{\link{teamBowlersVsBatsmenAllOppnAllMatchesPlot}}\cr
}

