% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBowlingScorecardMatch.R
\name{teamBowlingScorecardMatch}
\alias{teamBowlingScorecardMatch}
\title{Compute and return the bowling scorecard of a team in a match}
\usage{
teamBowlingScorecardMatch(match,theTeam)
}
\arguments{
\item{match}{The match between the teams}

\item{theTeam}{Team for which bowling performance is required}
}
\value{
l
A data frame with the bowling performance in alll matches against all oppositions
}
\description{
This function computes and returns the bowling scorecard of a team in a match
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#Get all matches between India  and other opposition
a <- getMatchDetails("England","Pakistan","2006-09-05",dir="../temp")

teamBowlingScorecardMatch(a,'England')
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{teamBowlingWicketMatch}}\cr
\code{\link{teamBowlersVsBatsmenMatch}}\cr
\code{\link{teamBattingScorecardMatch}}\cr
}
\author{
Tinniam V Ganesh
}
