% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamRunsAcrossAllOppnAllMatches.R
\name{teamRunsAcrossOversAllOppnAllMatches}
\alias{teamRunsAcrossOversAllOppnAllMatches}
\title{Compute the runs by team against all team in powerplay, middle and death overs}
\usage{
teamRunsAcrossOversAllOppnAllMatches(matches,t1,plot=1)
}
\arguments{
\item{t1}{The  team for which  the runs is required}

\item{plot}{Plot=1 (static), Plot=2(interactive)}

\item{match}{The dataframe of the match}
}
\value{
none
}
\description{
This function  plots the runs by team against all teams in in powerplay, middle and death overs
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{

# Plot tne match worm plot
teamRunsAcrossOversAllOppnAllMatches(matches,'England')
}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{getBatsmanDetails}}\cr
\code{\link{getBowlerWicketDetails}}\cr
\code{\link{batsmanDismissals}}\cr
\code{\link{getTeamBattingDetails}}\cr
}
\author{
Tinniam V Ganesh
}
