% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamRunSRDeathOversPlotMatch.R
\name{teamRunSRDeathOversPlotMatch}
\alias{teamRunSRDeathOversPlotMatch}
\title{Team batting plots runs vs SR in death overs for team  in match}
\usage{
teamRunSRDeathOversPlotMatch(match,t1, t2, plot=1)
}
\arguments{
\item{match}{Match}

\item{t1}{The team}

\item{t2}{The  opposition team}

\item{plot}{Plot=1 (static), Plot=2(interactive)}
}
\description{
This function computes and plots runs vs SR  in death overs of a team in match against
opposition.
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{


# Top batsman is displayed in descending order of runs
teamRunSRDeathOversPlotMatch(match,t1="India",t2="England")

}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{teamBatsmenVsBowlersAllOppnAllMatchesPlot}}\cr
\code{\link{teamBatsmenPartnershipOppnAllMatchesChart}}\cr
\code{\link{teamBatsmenPartnershipAllOppnAllMatchesPlot}}\cr
\code{\link{teamBowlingWicketRunsAllOppnAllMatches}}
}
\author{
Tinniam V Ganesh
}
