% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamRunSRDeathOversPlotOppnAllMatches.R
\name{teamRunSRDeathOversPlotOppnAllMatches}
\alias{teamRunSRDeathOversPlotOppnAllMatches}
\title{Team batting plots runs vs SR in death overs for team against opposition in all matches}
\usage{
teamRunSRDeathOversPlotOppnAllMatches(matches,t1, t2, plot=1)
}
\arguments{
\item{matches}{All matches of the team in all matches with all oppositions}

\item{t1}{The team}

\item{t2}{The  opposition team}

\item{plot}{Plot=1 (static), Plot=2(interactive)}
}
\description{
This function computes and plots runs vs SR  in middle overs of a team in all matches against
opposition.
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{


# Top batsman is displayed in descending order of runs
teamRunSRDeathOversPlotOppnAllMatches(matches,t1="India",t2="England")

}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{teamBatsmenVsBowlersAllOppnAllMatchesPlot}}\cr
\code{\link{teamBatsmenPartnershipOppnAllMatchesChart}}\cr
\code{\link{teamBatsmenPartnershipAllOppnAllMatchesPlot}}\cr
\code{\link{teamBowlingWicketRunsAllOppnAllMatches}}
}
\author{
Tinniam V Ganesh
}
