% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamRunsSRPlotOppnAllMatches.R
\name{teamRunsSRPlotOppnAllMatches}
\alias{teamRunsSRPlotOppnAllMatches}
\title{Team batting Runs vs SR  against  oppositions in all matches}
\usage{
teamRunsSRPlotOppnAllMatches(matches,t1,t2,plot=1)
}
\arguments{
\item{matches}{All matches of the team in all matches with opposition}

\item{t1}{The team t}

\item{t2}{The opposition team}

\item{plot}{Plot=1 (static), Plot=2(interactive)}
}
\description{
This function computes and plots the Runs vs SR  of a team in all matches against an
oppositions.
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{

# Top batsman is displayed in descending order of runs
teamRunsSRPlotOppnAllMatches(matches,theTeam="India")

# The best England players scorecard against India is shown
teamRunsSRPlotOppnAllMatches(matches,theTeam="England")
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{teamBatsmenVsBowlersAllOppnAllMatchesPlot}}\cr
\code{\link{teamBatsmenPartnershipOppnAllMatchesChart}}\cr
\code{\link{teamBatsmenPartnershipAllOppnAllMatchesPlot}}\cr
\code{\link{teamBowlingWicketRunsAllOppnAllMatches}}
}
\author{
Tinniam V Ganesh
}
