% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamRunsSRPowerPlayPlotAllOppnAllMatches.R
\name{teamRunsSRPowerPlayPlotAllOppnAllMatches}
\alias{teamRunsSRPowerPlayPlotAllOppnAllMatches}
\title{Team batting plots runs vs SR in powerplay for team against all oppositions in all matches}
\usage{
teamRunsSRPowerPlayPlotAllOppnAllMatches(matches,theTeam)
}
\arguments{
\item{matches}{All matches of the team in all matches with all oppositions}

\item{t1}{The team for which the the batting partnerships are sought


 @param plot
Plot=1 (static), Plot=2(interactive)}
}
\description{
This function computes and plots runs vs SR  in power play of a team in all matches against all
oppositions.
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{

# Top batsman is displayed in descending order of runs
teamRunsSRPowerPlayPlotAllOppnAllMatches(matches,t1="India",plot=1)

}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{teamBatsmenVsBowlersAllOppnAllMatchesPlot}}\cr
\code{\link{teamBatsmenPartnershipOppnAllMatchesChart}}\cr
\code{\link{teamBatsmenPartnershipAllOppnAllMatchesPlot}}\cr
\code{\link{teamBowlingWicketRunsAllOppnAllMatches}}
}
\author{
Tinniam V Ganesh
}
