% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamSRAcrossOvers.R
\name{teamSRAcrossOvers}
\alias{teamSRAcrossOvers}
\title{Compute the Strike Rate  in powerplay, middle and death overs}
\usage{
teamSRAcrossOvers(match,t1,t2,plot=1)
}
\arguments{
\item{match}{The dataframe of the match}

\item{t1}{The 1st team of the match}

\item{t2}{the 2nd team in the match}

\item{plot}{Plot=1 (static), Plot=2(interactive)}
}
\value{
none
}
\description{
This function  plots strike rate scored in powerplay, middle and death overs
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{

# Plot tne match worm plot
teamSRAcrossOvers(match,'England',"Pakistan")
}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{getBatsmanDetails}}\cr
\code{\link{getBowlerWicketDetails}}\cr
\code{\link{batsmanDismissals}}\cr
\code{\link{getTeamBattingDetails}}\cr
}
\author{
Tinniam V Ganesh
}
