% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamWicketsERDeathOversPlotMatch.R
\name{teamWicketsERDeathOversPlotMatch}
\alias{teamWicketsERDeathOversPlotMatch}
\title{Team wickets vs ER in death overs against opposition in match}
\usage{
teamWicketsERDeathOversPlotMatch(matches,t1, t2, plot=1)
}
\arguments{
\item{t1}{Team for which bowling performance is required}

\item{plot}{Plot=1 (static), Plot=2(interactive)}

\item{matches}{The matches of the team against all oppositions and all matches}
}
\description{
This function computes wickets vs ER in death overs against  oppositions in  match
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{teamBowlersVsBatsmenAllOppnAllMatchesMain}}\cr
\code{\link{teamBowlersVsBatsmenAllOppnAllMatchesPlot}}\cr
}
\author{
Tinniam V Ganesh
}
