% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamWicketsERPlotOppnAllMatches.R
\name{teamWicketsERPlotOppnAllMatches}
\alias{teamWicketsERPlotOppnAllMatches}
\title{Team wickets vs ER against  all opposition all matches}
\usage{
teamWicketsERPlotOppnAllMatches(matches,t1,t2,plot=1)
}
\arguments{
\item{matches}{The matches of the team against all oppositions and all matches}

\item{t1}{The 1st team of the match}

\item{t2}{the 2nd team in the match}

\item{plot}{Plot=1 (static), Plot=2(interactive)}
}
\description{
This function computes wickets vs ER against all oppositions in all matches
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{teamBowlersVsBatsmenAllOppnAllMatchesMain}}\cr
\code{\link{teamBowlersVsBatsmenAllOppnAllMatchesPlot}}\cr
}
\author{
Tinniam V Ganesh
}
