% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamWicketsERPowerPlayPlotMatch.R
\name{teamWicketsERPowerPlayPlotMatch}
\alias{teamWicketsERPowerPlayPlotMatch}
\title{Team wickets vs ER in powewrplay against   opposition in match}
\usage{
teamWicketsERPowerPlayPlotMatch(match,t1,t2, plot=1)
}
\arguments{
\item{match}{The match of the team against  opposition}

\item{t1}{Team for which bowling performance is required}

\item{t2}{Opposition Team}

\item{plot}{Plot=1 (static), Plot=2(interactive)}
}
\description{
This function computes wickets vs ER in powewrplay against  oppositions in  match
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
teamWicketsERPowerPlayPlotMatch(match,t1,t2,plot=1)
}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{teamBowlersVsBatsmenAllOppnAllMatchesMain}}\cr
\code{\link{teamBowlersVsBatsmenAllOppnAllMatchesPlot}}\cr
}
\author{
Tinniam V Ganesh
}
