% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topERBowlerAcrossOversAllOppnAllMatches.R
\name{topERBowlerAcrossOversAllOppnAllMatches}
\alias{topERBowlerAcrossOversAllOppnAllMatches}
\title{Compute the best ER by bowlers against all team in powerplay, middle and death overs}
\usage{
topERBowlerAcrossOversAllOppnAllMatches(matches,t1)
}
\arguments{
\item{t1}{The 1st team of the match}

\item{match}{The dataframe of the match}
}
\value{
none
}
\description{
This function  computes the best ER by bowlers against akk team in in powerplay, middle and death overs
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{

topERBowlerAcrossOversAllOppnAllMatches(matches,'England')
}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{getBatsmanDetails}}\cr
\code{\link{getBowlerWicketDetails}}\cr
\code{\link{batsmanDismissals}}\cr
\code{\link{getTeamBattingDetails}}\cr
}
\author{
Tinniam V Ganesh
}
