% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topERBowlerAcrossOversOppnAllMatches.R
\name{topERBowlerAcrossOversOppnAllMatches}
\alias{topERBowlerAcrossOversOppnAllMatches}
\title{Compute the best ER by bowlers against team in powerplay, middle and death overs}
\usage{
topERBowlerAcrossOversOppnAllMatches(matches,t1)
}
\arguments{
\item{matches}{The dataframe of the matches}

\item{t1}{The 1st team of the match}
}
\value{
none
}
\description{
This function  computes the best ER by bowlers against team in in powerplay, middle and death overs
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{


topERBowlerAcrossOversOppnAllMatches.R(matches,'England')
}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{getBatsmanDetails}}\cr
\code{\link{getBowlerWicketDetails}}\cr
\code{\link{batsmanDismissals}}\cr
\code{\link{getTeamBattingDetails}}\cr
}
\author{
Tinniam V Ganesh
}
