% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topSRBatsmenAcrossOversAllOppnAllMatches.R
\name{topSRBatsmenAcrossOversAllOppnAllMatches}
\alias{topSRBatsmenAcrossOversAllOppnAllMatches}
\title{Compute the highest SR by batsmen against all team in powerplay, middle and death overs}
\usage{
topSRBatsmenAcrossOversAllOppnAllMatches(matches,t1,plot=1)
}
\arguments{
\item{matches}{The dataframe of the matches}

\item{t1}{The team of the match}
}
\value{
none
}
\description{
This function  computes the highest SR by batsmen by batsman against all team in in powerplay, middle and death overs
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{

# Plot tne match worm plot
topSRBatsmenAcrossOversAllOppnAllMatches(matches,'England')
}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{getBatsmanDetails}}\cr
\code{\link{getBowlerWicketDetails}}\cr
\code{\link{batsmanDismissals}}\cr
\code{\link{getTeamBattingDetails}}\cr
}
\author{
Tinniam V Ganesh
}
