% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topWicketsBowlerAcrossOversOppnAllMatches.R
\name{topWicketsBowlerAcrossOversOppnAllMatches}
\alias{topWicketsBowlerAcrossOversOppnAllMatches}
\title{Compute the best ER by bowlers against team in powerplay, middle and death overs}
\usage{
topWicketsBowlerAcrossOversOppnAllMatches(matches,t1,plot=1)
}
\arguments{
\item{matches}{The dataframe of the matches}

\item{t1}{The 1st team of the match}

\item{plot}{Plot=1 (static), Plot=2(interactive)}
}
\value{
none
}
\description{
This function  computes the highest wickets by bowlers against team in in powerplay, middle and death overs
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{

# Plot tne match worm plot
topWicketsBowlerAcrossOversOppnAllMatches(matches,'England')
}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{getBatsmanDetails}}\cr
\code{\link{getBowlerWicketDetails}}\cr
\code{\link{batsmanDismissals}}\cr
\code{\link{getTeamBattingDetails}}\cr
}
\author{
Tinniam V Ganesh
}
