% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winProbLR.R
\name{winProbabilityLR}
\alias{winProbabilityLR}
\title{Plot the  win probability using Logistic Regression model}
\usage{
winProbabilityLR(match,t1,t2,plot=1)
}
\arguments{
\item{match}{The dataframe of the match}

\item{t1}{The 1st team of the match}

\item{t2}{the 2nd team in the match}

\item{plot}{Plot=1 (static), Plot=2(interactive)}
}
\value{
none
}
\description{
This function  plots the  win probability of the teams in a T20 match
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#Get the match details
a <- getMatchDetails("England","Pakistan","2006-09-05",dir="../temp")

# Plot tne match worm plot
winProbabilityLR(a,'England',"Pakistan")
}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{getBatsmanDetails}}\cr
\code{\link{getBowlerWicketDetails}}\cr
\code{\link{batsmanDismissals}}\cr
\code{\link{getTeamBattingDetails}}\cr
}
\author{
Tinniam V Ganesh
}
