% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batsmanCumulativeStrikeRate.R
\name{batsmanCumulativeStrikeRate}
\alias{batsmanCumulativeStrikeRate}
\title{Batsman's cumulative average strike rate}
\usage{
batsmanCumulativeStrikeRate(df,name= "A Leg Glance")
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of batsman}
}
\value{
None
}
\description{
This function computes and plots the cumulative average strike rate  of a batsman
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#Get the data frame for Kohli
kohli <- getBatsmanDetails(team="India",name="Kohli",dir=pathToFile)
batsmanCumulativeStrikeRate(kohli,"Kohli")
}
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{batsmanCumulativeAverageRuns}}
\code{\link{bowlerCumulativeAvgEconRate}}
\code{\link{bowlerCumulativeAvgWickets}}
\code{\link{batsmanRunsVsStrikeRate}}
\code{\link{batsmanRunsPredict}}
}
\author{
Tinniam V Ganesh
}
