% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batsmanDismissals.R
\name{batsmanDismissals}
\alias{batsmanDismissals}
\title{Dismissal type of batsmen}
\usage{
batsmanDismissals(df,name="A Leg Glance",dateRange)
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of batsman}

\item{dateRange}{Date interval to consider}
}
\value{
None
}
\description{
This function computes and plots the type of dismissals of the
the batsman
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#Get the data frame for Kohli
kohli <- getBatsmanDetails(team="India",name="Kohli",dir=pathToFile)
batsmanDismissals(kohli,"Kohli",dateRange)
}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{batsmanFoursSixes}}\cr
\code{\link{batsmanRunsVsDeliveries}}\cr
\code{\link{batsmanRunsVsStrikeRate}}\cr
}
\author{
Tinniam V Ganesh
}
