% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batsmanFoursSixes.R
\name{batsmanFoursSixes}
\alias{batsmanFoursSixes}
\title{Batsman's total runs, fours and sixes}
\usage{
batsmanFoursSixes(df,name= "A Leg Glance")
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of batsman}
}
\value{
None
}
\description{
This function computes and plots the total runs, fours and sixes of
the batsman
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#Get the data frame for Kohli
kohli <- getBatsmanDetails(team="India",name="Kohli",dir=pathToFile)
kohli46 <- select(kohli,batsman,ballsPlayed,fours,sixes,runs)
batsmanFoursSixes(kohli46,"Kohli")
}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{batsmanDismissals}}
\code{\link{batsmanRunsVsDeliveries}}
\code{\link{batsmanRunsVsStrikeRate}}
\code{\link{batsmanRunsVsStrikeRate}}
\code{\link{batsmanRunsPredict}}
}
\author{
Tinniam V Ganesh
}
