% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batsmansRunsVsDeliveries.R
\name{batsmanRunsVsDeliveries}
\alias{batsmanRunsVsDeliveries}
\title{Plot the runs scored versus the deliveries. Fit a regression line}
\usage{
batsmanRunsVsDeliveries(df, name= "A Late Cut")
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of batsman}
}
\value{
None
}
\description{
This function plots the runs scored and the deliveries required. A  regression
smoothing function is used to fit the points
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get the data frame for Kohli
kohli <- getBatsmanDetails(team="India",name="Kohli")
batsmanRunsVsDeliveries(kohli,"Kohli")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{batsmanFoursSixes}}
\code{\link{batsmanRunsVsDeliveries}}
\code{\link{batsmanRunsVsStrikeRate}}
}

