% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batsmanRunsVsStrikeRate.R
\name{batsmanRunsVsStrikeRate}
\alias{batsmanRunsVsStrikeRate}
\title{Compute and plot the strike rate of the batsman versus the runs scored}
\usage{
batsmanRunsVsStrikeRate(kohli,name="Kohli")
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of batsman}
}
\value{
None
}
\description{
This function plots the runs scored by the batsman  and the runs scored
by the batsman. A loess line is fitted over the points
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get the data frame for Kohli
kohli <- getBatsmanDetails(team="India",name="Kohli")
batsmanRunsVsStrikeRate(kohli,"Kohli")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{batsmanDismissals}}
\code{\link{batsmanRunsVsDeliveries}}
\code{\link{batsmanRunsVsStrikeRate}}
\code{\link{batsmanRunsVsStrikeRate}}
\code{\link{batsmansRunsPredict}}
\code{\link{teamBatsmanPartnershipAllOppnAllMatches}}
}

