% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlerMeanRunsConceded.R
\name{bowlerMeanRunsConceded}
\alias{bowlerMeanRunsConceded}
\title{Compute and plot mean runs  conceded by the bowler}
\usage{
bowlerMeanRunsConceded(df, name)
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of bowler}
}
\value{
None
}
\description{
This function computes and plots mean runs conceded by the bowler for the
number of overs bowled by the bowler
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get the data frame for RA Jadeja
jadeja <- getBowlerWicketDetails(team="India",name="Jadeja")
bowlerMeanRunsConceded(jadeja,"RA Jadeja")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{bowlerMovingAverage}}
\code{\link{bowlerWicketPlot}}
\code{\link{bowlerWicketsVenue}}
\code{\link{bowlerMeanRunsConceded}}
}

