% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlerWicketPlot.R
\name{bowlerWicketPlot}
\alias{bowlerWicketPlot}
\title{Compute and plot average wickets taken versus the number of overs bowled}
\usage{
bowlerWicketPlot(df, name)
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of bowler}
}
\value{
None
}
\description{
This function computes and plots the average wickets taken by the bowler versus the
number of overs bowled
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get the data frame for RA Jadeja
jadeja <- getBowlerWicketDetails(team="India",name="Jadeja")
bowlerWicketPlot(jadeja,"RA Jadeja")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{bowlerMeanEconomyRate}}
\code{\link{bowlerWicketsVenue}}
\code{\link{bowlerMeanRunsConceded}}
}

