% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlerWicketPlot.R
\name{bowlerWicketPlot}
\alias{bowlerWicketPlot}
\title{Average wickets versus of overs bowled}
\usage{
bowlerWicketPlot(df, name)
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of bowler}
}
\value{
None
}
\description{
This function computes and plots the average wickets taken by the bowler versus the
number of overs bowled
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
# Get the data frame for RA Jadeja
jadeja <- getBowlerWicketDetails(team="India",name="Jadeja",dir=pathToFile)
bowlerWicketPlot(jadeja,"RA Jadeja")
}
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{bowlerMeanEconomyRate}}\cr
\code{\link{bowlerWicketsVenue}}\cr
\code{\link{bowlerMeanRunsConceded}}\cr
}
\author{
Tinniam V Ganesh
}
