% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowlerWicketsAgainstOppn.R
\name{bowlerWicketsAgainstOpposition}
\alias{bowlerWicketsAgainstOpposition}
\title{Bowler wickets versus different teams}
\usage{
bowlerWicketsAgainstOpposition(df, name)
}
\arguments{
\item{df}{Data frame}

\item{name}{Name of bowler}
}
\value{
None
}
\description{
This function computes and plots mean number of wickets taken by the bowler  against different
opposition
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
# Get the data frame for RA Jadeja
jadeja <- getBowlerWicketDetails(team="India",name="Jadeja",dir=pathToFile)
bowlerWicketsAgainstOpposition(jadeja,"RA Jadeja")
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{bowlerMovingAverage}}\cr
\code{\link{bowlerWicketPlot}}\cr
\code{\link{bowlerWicketsVenue}}\cr
\code{\link{bowlerMeanRunsConceded}}\cr
}
\author{
Tinniam V Ganesh
}
