% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBatsmanDetails.R
\name{getBatsmanDetails}
\alias{getBatsmanDetails}
\title{Get batting details of  batsman from match}
\usage{
getBatsmanDetails(team,name,dir=".")
}
\arguments{
\item{team}{The team of the batsman e.g. India}

\item{name}{Name of batsman}

\item{dir}{The directory where the source file exists}
}
\value{
None
}
\description{
This function gets the batting details of a batsman given the match data as a RData file
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
getBatsmanDetails(team="India",name="Kohli",dir=pathToFile)
}


}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{batsmanRunsPredict}}\cr
\code{\link{batsmanMovingAverage}}\cr
\code{\link{bowlerWicketsVenue}}\cr
\code{\link{bowlerMeanRunsConceded}}\cr
}
\author{
Tinniam V Ganesh
}
