% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBowlerWicketDetails.R
\name{getBowlerWicketDetails}
\alias{getBowlerWicketDetails}
\title{Get the bowling details of a bowler of a team}
\usage{
getBowlerWicketDetails(team,name)
}
\arguments{
\item{team}{The tam to which the bowler belongs}

\item{name}{The name of the bowler}
}
\value{
dataframe
The dataframe of bowling performance
}
\description{
This function gets the bowling of a bowler (overs,maidens,runs,wickets,venue, opposition)
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get the bowling details of bowlers of a team e.g. India. This is saved as a dataframe
c <- getTeamBowlingDetails("India",dir="../data",save=TRUE)
#Get the bowler details from this overall data frame
jadeja <- getBowlerWicketDetails(team="India",name="Jadeja")

# The dataframe from the above call is used in many functions
#bowlerMeanEconomyRate(jadeja,"RA Jadeja")

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{bowlerMovingAverage}}
\code{\link{getTeamBowlingDetails}}
\code{\link{bowlerMeanRunsConceded}}
\code{\link{teamBowlersWicketRunsOppnAllMatches}}
}

