% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTeamBattingDetails.R
\name{getTeamBattingDetails}
\alias{getTeamBattingDetails}
\title{Get teh team batting details against all oppositions in all matches}
\usage{
getTeamBattingDetails(team,dir=".",save=FALSE)
}
\arguments{
\item{team}{The team for which all matches against all oppositions is required}

\item{dir}{The source directory of RData files obtained with  convertAllYaml2RDataframes()}

\item{save}{Whether the data frame needs to be saved as RData or not. It is recommended to set save=TRUE
as the data can be used for a lot of analyses of batsmen}
}
\value{
battingDetails
The dataframe with the batting details
}
\description{
This function  gets the batting details of a team in all matchs against all
oppositions. This gets all the details of the batsmen balls faced,4s,6s,strikerate, runs, venue etc.
This function is then used for analyses of batsmen. This function calls teamBattingPerfDetails()
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
a <- getTeamBattingDetails("India",dir="../data", save=TRUE)

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{getBatsmanDetails}}\cr
\code{\link{getBowlerWicketDetails}}\cr
\code{\link{batsmanDismissals}}\cr
\code{\link{batsmanFoursSixes}}
}

