% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper}
\alias{helper}
\title{Gets min,max date and min and max matches from dataframe}
\usage{
helper(teamNames,odir=".")
}
\arguments{
\item{teamNames}{The team names}

\item{odir}{The output directory}
}
\value{
minDate,maxDate, minMatches, maxMatches
}
\description{
This function gets min,max date and min and max matches from dataframe
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Batsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
