% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper3.R
\name{helper3}
\alias{helper3}
\title{Gets min and max date from dataframe}
\usage{
helper3(teamNames,yearValue,odir=".")
}
\arguments{
\item{teamNames}{The team names}

\item{odir}{The output directory}
}
\value{
minDate,maxDate, bowingDF
}
\description{
This function creates a single datframe of all IPL batsmen and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#
iplBatsmanRank <- rankIPLBatsmen()
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankIPLBowlers}}\cr
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Batsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
