% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper3.R
\name{helper3}
\alias{helper3}
\title{Gets min,max date and min and max matches from dataframe for the year}
\usage{
helper3(teamNames,dateRange, dir=".",type="IPL")
}
\arguments{
\item{teamNames}{The team names}

\item{dateRange}{Date interval to consider}

\item{dir}{The input directory}

\item{type}{T20 format}
}
\value{
minDate,maxDate, bowingDF
}
\description{
This function gets min,max date and min and max matches from dataframe
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Batsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
