% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallRunsSRPlotT20M.R
\name{overallRunsSRPlotT20M}
\alias{overallRunsSRPlotT20M}
\title{Plot the Runs vs SR of Intl. T20 batsmen}
\usage{
overallRunsSRPlotT20M(dir=".", dateRange)
}
\arguments{
\item{dateRange}{Date interval to consider}

\item{teamNames}{The team names}

\item{odir}{The output directory}

\item{minMatches}{Minimum matches played}

\item{runsvsSR}{Runs or Strike rate}
}
\value{
The ranked T20 batsmen
}
\description{
This function creates a single datframe of all T20 batsmen and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
rankT20Batsmen(dir=".",dateRange)
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
