% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallWicketsERPlotT20.R
\name{overallWicketsERPlotT20}
\alias{overallWicketsERPlotT20}
\title{Ranks the T20 bowlers}
\usage{
overallWicketsERPlotT20(dir=".",minMatches, dateRange,type="IPL",plot=1)
}
\arguments{
\item{dir}{The input directory}

\item{minMatches}{Minimum matches played}

\item{dateRange}{Date interval to consider}

\item{type}{T20 league}

\item{plot}{plot=1 (static),plot=2(interactive), plot=3 (table)}
}
\description{
This function plots the overall wickets vs ER in T20 men
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
overallWicketsERPlotT20(dir=".",minMatches, dateRange,type="IPL",plot=1)
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Batsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
