% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseYamlOver.R
\name{parseYamlOver}
\alias{parseYamlOver}
\title{Parse the yaml file and convert to data frame (This is an internal function)}
\usage{
parseYamlOver(match,s,ateam,overs,delivery,meta)
}
\arguments{
\item{match}{The dataframe of the match}

\item{s}{The string with the delivery}

\item{ateam}{The team}

\item{overs}{overs}

\item{delivery}{The delivery of the over}

\item{meta}{The meta information of the match}
}
\value{
overs
The dataframe of overs
}
\description{
This function parses the yaml file and converts it into a data frame. This is an internal function and
is used by convertAllYaml2RDataframes() & convertYaml2RDataframe()
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Parse the yaml over


}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{getBatsmanDetails}}\cr
\code{\link{getBowlerWicketDetails}}\cr
\code{\link{batsmanDismissals}}\cr
\code{\link{getTeamBattingDetails}}
}

