% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWinLossBetweenTeams.R
\name{plotWinLossBetweenTeams}
\alias{plotWinLossBetweenTeams}
\title{Plot  wins for each team}
\usage{
plotWinLossBetweenTeams(team1,team2,dir=".")
}
\arguments{
\item{team1}{The 1st team}

\item{team2}{The 2nd team}

\item{dir}{The source directory of teh RData files}
}
\value{
None
}
\description{
This function computes and plots number of wins for each team in all their
encounters. The plot includes the number of  wins byteam1 each team and the matches
with no result
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{

plotWinLossBetweenTeams(team1="India",team2="Australia",dir=pathToFile)
batsmanDismissals(kohli,"Kohli")
}
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{batsmanFoursSixes}}\cr
\code{\link{batsmanRunsVsDeliveries}}\cr
\code{\link{batsmanRunsVsStrikeRate}}\cr
}
\author{
Tinniam V Ganesh
}
