% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankIPLBatsmen.R
\name{rankIPLBatsmen}
\alias{rankIPLBatsmen}
\title{Ranks the IPL batsmen}
\usage{
rankIPLBatsmen()
}
\value{
The ranked IPL batsmen
}
\description{
This function creates a single datframe of all IPL batsmen and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#
iplBatsmanRank <- rankIPLBatsmen()
}

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}\cr
\url{https://github.com/tvganesh/yorkrData}
}
\seealso{
\code{\link{rankIPLBowlers}}\cr
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Batsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}

