% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankIPLBowlers.R
\name{rankIPLBowlers}
\alias{rankIPLBowlers}
\title{Ranks the IPL bowlers}
\usage{
rankIPLBowlers(dir='.',odir=".",minMatches=20)
}
\arguments{
\item{odir}{The output directory}

\item{minMatches}{Minimum matches}

\item{dir}{The input directory}
}
\value{
The ranked IPL batsmen
}
\description{
This function creates a single datframe of all IPL bowlers and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#
iplBowlersRank <- rankIPLBowlers()
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankIPLBatsmen}}\cr
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Batsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
