% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankT20Bowlers.R
\name{rankT20Bowlers}
\alias{rankT20Bowlers}
\title{Ranks the T20 bowlers}
\usage{
rankT20Bowlers(teamNames,odir=".",minMatches, yearSelected, wicketsVsER)
}
\arguments{
\item{teamNames}{The team names}

\item{odir}{The output directory}

\item{minMatches}{Minimum matches played}

\item{yearSelected}{Selected year}

\item{wicketsVsER}{Wickets  or economy rate}
}
\value{
The ranked T20 bowlers
}
\description{
This function creates a single datframe of all T20 bowlers and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
rankT20Bowlers(teamNames,odir=".",minMatches, yearSelected, wicketsVsER)
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Batsmen}}\cr
\code{\link{rankT20Bowlers}}\cr
}
\author{
Tinniam V Ganesh
}
