% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankT20Bowlers.R
\name{rankT20Bowlers}
\alias{rankT20Bowlers}
\title{Ranks the T20 bowlers}
\usage{
rankT20Bowlers(dir='.',odir=".",minMatches=20)
}
\arguments{
\item{dir}{The input directory}

\item{odir}{The output directory}

\item{minMatches}{Minimum matches}
}
\value{
The ranked T20 bowlers
}
\description{
This function creates a single datframe of all T20 batsmen and then ranks them
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#
t20BowlersRank <- rankT20Bowlers()
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{rankIPLBowlers}}\cr
\code{\link{rankODIBowlers}}\cr
\code{\link{rankODIBatsmen}}\cr
\code{\link{rankT20Batsmen}}\cr
}
\author{
Tinniam V Ganesh
}
