% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAllMatchesBetweenTeams.R
\name{saveAllMatchesBetweenTeams}
\alias{saveAllMatchesBetweenTeams}
\title{This function saves all matches between 2 teams as a single dataframe}
\usage{
saveAllMatchesBetweenTeams()
}
\value{
None
}
\description{
This function saves all matches between 2 teams as a single dataframe in the
current directory
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
saveAllMatchesBetweenTeams
}
}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{batsmanDismissals}}
\code{\link{batsmanRunsVsDeliveries}}
\code{\link{batsmanRunsVsStrikeRate}}
\code{\link{getAllMatchesAllOpposition}}
\code{\link{getAllMatchesBetweenTeams}}
}

