% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialProc.R
\name{specialProc}
\alias{specialProc}
\title{This function is used to parse yaml file where deliveries are bigger than 10 deliveries (This is an internal
function)}
\usage{
specialProc(dist, overset, ateam,over,str1,meta)
}
\arguments{
\item{dist}{}

\item{overset}{}

\item{ateam}{The team}

\item{over}{over}

\item{str1}{}

\item{meta}{The meta information of the match}
}
\value{
over
The dataframe of over
}
\description{
This is special processing function. This is an internal function and
is used by convertAllYaml2RDataframes() & convertYaml2RDataframe()
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Parse the yaml over


}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{getBatsmanDetails}}\cr
\code{\link{getBowlerWicketDetails}}\cr
\code{\link{batsmanDismissals}}\cr
\code{\link{getTeamBattingDetails}}
}

