% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBatsmenPartnershipOppnAllMatches.R
\name{teamBatsmenPartnershiOppnAllMatches}
\alias{teamBatsmenPartnershiOppnAllMatches}
\title{Team batting partnership against a opposition all matches}
\usage{
teamBatsmenPartnershiOppnAllMatches(matches,theTeam,report="summary")
}
\arguments{
\item{matches}{All the matches of the team against the oppositions}

\item{theTeam}{The team for which the the batting partnerships are sought}

\item{report}{If the report="summary" then the list of top batsmen with the highest partnerships is displayed. If
report="detailed" then the detailed break up of partnership is returned as a dataframe}
}
\value{
partnerships
The data frame of the partnerships
}
\description{
This function computes the performance of batsmen against all bowlers of an oppositions in all matches. This
function returns a dataframe
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
# Get all matches for team India against all oppositions
matches <- getAllMatchesBetweenTeams("Australia","India",dir="../data")
# You can also directly load the data
#load("India-Australia-allMatches.RData")

m <-teamBatsmenPartnershiOppnAllMatches(a,'India',report="summary")
m <-teamBatsmenPartnershiOppnAllMatches(a,'Australia',report="detailed")
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{teamBatsmenVsBowlersAllOppnAllMatchesPlot}}\cr
\code{\link{teamBatsmenPartnershipOppnAllMatchesChart}}\cr
\code{\link{teamBowlersVsBatsmenMatch}}\cr
\code{\link{teamBattingScorecardMatch}}\cr
}
\author{
Tinniam V Ganesh
}
