% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBatsmenPartnershipMatch.R
\name{teamBatsmenPartnershipMatch}
\alias{teamBatsmenPartnershipMatch}
\title{This function plots the partnerships of batsmen in a match against an opposition. It can also return
the data frame}
\usage{
teamBatsmenPartnershipMatch(match,theTeam,plot=TRUE)
}
\arguments{
\item{match}{The match between the teams}

\item{theTeam}{The team for which the the batting partnerships are sought}

\item{plot}{If plot=TRUE then a plot is created otherwise a data frame is returned}
}
\value{
df
The data frame of the batsmen partnetships
}
\description{
This function plots the partnerships of batsmen in a match against an opposition or it can return
the data frame
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
# Get athe match between England and Pakistan
a <- getMatchDetails("England","Pakistan","2006-09-05",dir="../temp")
batsmenPartnershipMatch(a,"Pakistan")
batsmenPartnershipMatch(a,"England",plot=TRUE)
m <-batsmenPartnershipMatch(a,"Pakistan",plot=FALSE)

}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{teamBatsmenVsBowlersAllOppnAllMatchesPlot}}
\code{\link{teamBatsmenPartnershipOppnAllMatchesChart}}
\code{\link{teamBatsmenPartnershipAllOppnAllMatchesPlot}}
\code{\link{teamBatsmenVsBowlersOppnAllMatches}}
}

