% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBatsmenVsBowlersMatch.R
\name{teamBatsmenVsBowlersMatch}
\alias{teamBatsmenVsBowlersMatch}
\title{Team batsmen against bowlers in a match}
\usage{
teamBatsmenVsBowlersMatch(match,theTeam,opposition, plot=1)
}
\arguments{
\item{match}{The match between the teams}

\item{theTeam}{The team for which the the batting partnerships are sought}

\item{opposition}{The opposition team}

\item{plot}{lot=1 (static),plot=2(interactive),plot=3(table)}
}
\value{
b
The data frame of the batsmen vs bowlers performance
}
\description{
This function plots the performance of  batsmen versus bowlers in a match  or it can return
the data frame
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
# Get athe match between England and Pakistan
a <- getMatchDetails("England","Pakistan","2006-09-05",dir="../temp")
batsmenVsBowlersMatch(a,'Pakistan','England', plot=TRUE)
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{teamBatsmenVsBowlersAllOppnAllMatchesPlot}}\cr
\code{\link{teamBatsmenPartnershipOppnAllMatchesChart}}\cr
\code{\link{teamBatsmenPartnershipAllOppnAllMatchesPlot}}\cr
\code{\link{teamBattingScorecardMatch}}\cr
}
\author{
Tinniam V Ganesh
}
