% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBattingPerfDetails.R
\name{teamBattingPerfDetails}
\alias{teamBattingPerfDetails}
\title{Gets the team batting details.}
\usage{
teamBattingPerfDetails(match,theTeam,includeInfo=FALSE)
}
\arguments{
\item{match}{The match between the teams}

\item{theTeam}{The team for which the the batting partnerships are sought}

\item{includeInfo}{Whether to include venue,date, winner and result}
}
\value{
df
dataframe
}
\description{
This function gets the team batting detals
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#teamBattingPerfDetails()
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{teamBatsmenVsBowlersAllOppnAllMatchesPlot}}\cr
\code{\link{teamBatsmenPartnershipOppnAllMatchesChart}}\cr
\code{\link{teamBatsmenPartnershipAllOppnAllMatchesPlot}}\cr
\code{\link{teamBattingScorecardMatch}}\cr
}
\author{
Tinniam V Ganesh
}
