% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBattingScorecardMatch.R
\name{teamBattingScorecardMatch}
\alias{teamBattingScorecardMatch}
\title{Compute and return the batting scorecard of a team in a match}
\usage{
teamBattingScorecardMatch(match,theTeam)
}
\arguments{
\item{match}{The match for which the score card is required e.g.}

\item{name}{Team for which scorecard required}
}
\value{
scorecard
A data frame with the batting scorecard
}
\description{
This function computes returns the batting scorecard (runs, fours, sixes, balls played) for the
team
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
a <- getMatchDetails("England","Pakistan","2006-09-05",dir="../temp")
teamBowlingScorecardMatch(a,'England')


}
\author{
Tinniam V Ganesh
}
\references{
\url{http://cricsheet.org/}\cr
\url{https://gigadom.wordpress.com/}
}
\seealso{
\code{\link{teamBatsmenPartnershipAllOppnAllMatches}}
\code{\link{teamBatsmenPartnershipAllOppnAllMatchesPlot}}
\code{\link{teamBatsmenPartnershipOppnAllMatchesChart}}
}

