% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teamBowlingWicketKindMatch.R
\name{teamBowlingWicketKindMatch}
\alias{teamBowlingWicketKindMatch}
\title{Compute and plot the wicket kinds by bowlers in match}
\usage{
teamBowlingWicketKindMatch(match,theTeam,opposition,plot=1)
}
\arguments{
\item{match}{The match between the teams}

\item{theTeam}{Team for which bowling performance is required}

\item{opposition}{The opposition team}

\item{plot}{plot=1 (static),plot=2(interactive),plot=3(table)}
}
\value{
None or data fame
A data frame with the bowling performance in alll matches against all oppositions
}
\description{
This function computes returns kind of wickets (caught, bowled etc) of bowlers in a match between 2 teams
}
\note{
Maintainer: Tinniam V Ganesh \email{tvganesh.85@gmail.com}
}
\examples{
\dontrun{
#Get the match details
a <- getMatchDetails("England","Pakistan","2006-09-05",dir="../temp")

teamBowlingWicketKindMatch(a,"England","Pakistan",plot=FALSE)
teamBowlingWicketKindMatch(a,"Pakistan","England")
}

}
\references{
\url{https://cricsheet.org/}\cr
\url{https://gigadom.in/}\cr
\url{https://github.com/tvganesh/yorkrData/}
}
\seealso{
\code{\link{teamBowlingWicketMatch}}\cr
\code{\link{teamBowlingWicketRunsMatch}}\cr
\code{\link{teamBowlersVsBatsmenMatch}}\cr
}
\author{
Tinniam V Ganesh
}
